% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_teams.R
\name{rank_teams}
\alias{rank_teams}
\title{Classify tRivia teams based on score}
\usage{
rank_teams(res)
}
\arguments{
\item{res}{data frame containing at least two colums: `team`, `score` and `round` (optional)}
}
\description{
Classify tRivia teams based on score
}
