% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_trivia.R
\name{score_trivia}
\alias{score_trivia}
\title{Calculate scores for a tRivia team based on a data.frame with correct answers}
\usage{
score_trivia(data, n_round = NULL, score, ref)
}
\arguments{
\item{data}{url or a data frame with the results of a multiple choice google form pool}

\item{n_round}{vector indicating the round of questions, if applies}

\item{score}{Numeric value indicating the score for questions at the round}

\item{ref}{Data frame containing two columns: 'question_name' and 'correct_answer'}
}
\description{
Calculate scores for a tRivia team based on a data.frame with correct answers
}
